section7

+++ resources/item.py
@@ -1,6 +1,12 @@
 from flask import request
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import (
+    jwt_required,
+    get_jwt_claims,
+    jwt_optional,
+    get_jwt_identity,
+    fresh_jwt_required
+)
 
 from models.item import ItemModel
 
@@ -21,7 +27,7 @@ class Item(Resource):
         )
         return parser.parse_args()
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
 
@@ -29,6 +35,7 @@ class Item(Resource):
             return item.json()
         return {'message': 'item not found'}, 404
 
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': 'El item de nombre {} ya existe'.format(name)}, 400
@@ -59,15 +66,26 @@ class Item(Resource):
         
         return item.json(), 201
 
-
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin requered!'}, 401 
         item = ItemModel.find_by_name(name)
         if item:
             item.delete()
-        return {'message': 'Item delete'}
+            return {'message': 'Item delete'}
+        return {'message': 'Item not exist'}, 404
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {'items': [i.json() for i in ItemModel.query.all()]}
-        #whith lambda
-        #return {'items': list(map(lambda x: x.json(), ItemModel.query.all()))}
+        #como jwt es opcional, si no esta login el get da None
+        user_id = get_jwt_identity()
+        items = [i.json() for i in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}, 200
+        return {
+            'items': [i['name'] for i in items],
+            'message': 'More data availlabel if yo log in'
+        }, 200
